compare_Product/requirements.txt
django-froala-editor==4.0.19
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
phonenumbers==8.13.15
Pillow==9.5.0
